Move styles outside of the pages directory
