+++ pages/people.js
@@ -2,8 +2,8 @@ import React from "react";
 import Link from "next/link";
 import Layout from "../components/Layout";
 import sanity from "../lib/sanity";
-import styles from "./styles/people";
-import listStyles from "./styles/list";
+import styles from "../styles/people";
+import listStyles from "../styles/list";
 import imageUrlBuilder from "@sanity/image-url";
 import sanityClient from "../lib/sanity";
 const imageBuilder = imageUrlBuilder(sanityClient);
