+++ styles/people.js
